requirements.txt
alembic==1.10.4
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.3
Deprecated==1.2.14
dnspython==2.4.2
dparse==0.6.3
et-xmlfile==1.1.0
Flask==2.3.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-oauth==1.0.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.5.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mdurl==0.1.2
numpy==1.24.2
oauthlib==3.2.2
openpyxl==3.1.2
ordered-set==4.1.0
packaging==21.3
pandas==2.0.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
Pygments==2.16.1
PyJWT==2.8.0
pymongo==4.4.1
PyMySQL==1.0.3
pyOpenSSL==23.2.0
pyotp==2.8.0
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-magic==0.4.27
python-socketio==5.8.0
pytz==2023.3
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safety==2.3.5
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.4
wrapt==1.15.0
WTForms==3.0.1
zipp==3.15.0
